pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roadmesh"
version = "0.1.0"
description = "ML pipeline for road mesh extraction from satellite imagery"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Azat Foxie", email = "your-email@example.com"}
]
keywords = ["machine-learning", "road-extraction", "satellite-imagery", "gis", "mesh"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "segmentation-models-pytorch>=0.3.3",
    "albumentations>=1.3.1",
    "timm>=0.9.12",
    
    # Data processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "geopandas>=0.14.0",
    "rasterio>=1.3.9",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "mercantile>=1.2.1",
    
    # Image processing
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "scikit-image>=0.21.0",
    
    # API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.6",
    
    # Configuration & Logging
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    
    # Utilities
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "diskcache>=5.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "ipykernel>=6.27.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
]

train = [
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
    "onnx>=1.15.0",
    "onnxruntime-gpu>=1.16.0",
]

[project.scripts]
roadmesh = "roadmesh.cli:main"

[project.urls]
Homepage = "https://github.com/your-username/roadmesh"
Documentation = "https://github.com/your-username/roadmesh#readme"
Repository = "https://github.com/your-username/roadmesh"
Issues = "https://github.com/your-username/roadmesh/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
roadmesh = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | data
)/
'''

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "data",
]

[tool.ruff.isort]
known-first-party = ["roadmesh"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "albumentations.*",
    "segmentation_models_pytorch.*",
    "rasterio.*",
    "mercantile.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=roadmesh --cov-report=term-missing"
